Add GigaChatAPI interface
